refactor: remove obsolete deleteAccounts
